Feature Added: Added Population Sampling

+++ Calculator/Addition.py
@@ -1,5 +1,5 @@
 def addition(a, b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     c = a + b
     return c
